test(auth): transport test files を責務別に分割し 200 行制限を遵守
Codex Round 5 で confidence 100 の規約違反指摘:
- apps/frontend/src/auth/webAuthTransport.test.ts (276 行)
- apps/mobile/src/auth/mobileAuthTransport.test.ts (448 行)

責務別に分割しヘルパーを共通モジュール化:

# Web
- _webAuthTransportTestHelpers.ts (56 行): makeTransport / createTokenHolder /
  jsonResponse / emptyResponse / validSessionBody / apiUrl の共通 helper
- webAuthTransport.test.ts (159 行): refreshSession + login + state
- webAuthTransport.logout.test.ts (67 行): logout のみ (5 cases)

# Mobile
- _mobileAuthTransportTestHelpers.ts (57 行): 同上 + REFRESH_TOKEN_KEY
- mobileAuthTransport.test.ts (131 行): refreshSession (7 cases)
- mobileAuthTransport.flows.test.ts (155 行): login + setAccessToken /
  persistSession / clearPersistedSession (8 cases)
- mobileAuthTransport.logout.test.ts (156 行): logout (6 cases including
  401→refresh→retry rotation, expired stuck 回避, transient retry 可)

vi.mock は hoisting の制約で各 test ファイル先頭に保持する必要があるため
shared module には移さない (pure helpers のみ共通化)。

検証:
- pnpm run test-once: 184 files / 2132 tests pass
- pnpm run tsc: pass
