+++ apps/frontend/src/auth/webAuthTransport.test.ts
@@ -7,64 +7,14 @@ vi.mock("@packages/sync-engine", () => ({
   trackServerTimeFromResponse: vi.fn(),
 }));
 
-import { createWebAuthTransport } from "./webAuthTransport";
-
-const apiUrl = "http://localhost:3456";
-
-// 既存テストは vi.stubGlobal("fetch", ...) で global fetch を mock する前提なので、
-// authenticatedFetch も global fetch にデリゲートする実装をデフォルトにする。
-// logout のテストだけは authenticatedFetch を直接 mock したいので opts で override
-function makeTransport(opts?: {
-  tokenHolder?: {
-    getToken: () => string | null;
-    setToken: (t: string | null) => void;
-  };
-  authenticatedFetch?: typeof fetch;
-}) {
-  const authenticatedFetch =
-    opts?.authenticatedFetch ?? ((input, init) => fetch(input, init));
-  return createWebAuthTransport(
-    { apiUrl, authenticatedFetch },
-    opts?.tokenHolder ?? createTokenHolder(),
-  );
-}
-
-function createTokenHolder() {
-  let token: string | null = null;
-  return {
-    getToken: () => token,
-    setToken: (t: string | null) => {
-      token = t;
-    },
-  };
-}
-
-function jsonResponse(body: unknown, status = 200): Response {
-  return new Response(JSON.stringify(body), {
-    status,
-    headers: { "Content-Type": "application/json" },
-  });
-}
-
-function emptyResponse(status: number): Response {
-  return new Response(null, { status });
-}
-
-function validSessionBody(token = "jwt-token") {
-  return {
-    token,
-    user: {
-      id: "user-1",
-      name: "Test",
-      providers: [],
-      plan: "free",
-      tabPreference: {
-        tabs: ["home", "daily", "stats", "goals", "tasks"],
-        updatedAt: "2026-05-14T10:00:00.000Z",
-      },
-    },
-  };
-}
+import {
+  apiUrl,
+  createTokenHolder,
+  emptyResponse,
+  jsonResponse,
+  makeTransport,
+  validSessionBody,
+} from "./_webAuthTransportTestHelpers";
 
 beforeEach(() => {
   vi.clearAllMocks();
@@ -91,14 +41,18 @@ describe("webAuthTransport", () => {
         expect(result.session.token).toBe("new-jwt");
         expect(result.session.user.id).toBe("user-1");
       }
-      // credentials: include + /auth/token に POST
-      const [, init] = fetchMock.mock.calls[0];
-      expect(fetchMock.mock.calls[0][0]).toBe(`${apiUrl}/auth/token`);
-      expect((init as RequestInit).method).toBe("POST");
-      expect((init as RequestInit).credentials).toBe("include");
+
+      // POST /auth/token + credentials: include
+      expect(fetchMock).toHaveBeenCalledWith(
+        `${apiUrl}/auth/token`,
+        expect.objectContaining({
+          method: "POST",
+          credentials: "include",
+        }),
+      );
     });
 
-    it("401 -> { kind: 'expired' }", async () => {
+    it("401 -> { kind: 'expired' } (session 復元不能)", async () => {
       vi.stubGlobal("fetch", vi.fn().mockResolvedValue(emptyResponse(401)));
       const transport = makeTransport();
 
@@ -106,125 +60,55 @@ describe("webAuthTransport", () => {
       expect(result.kind).toBe("expired");
     });
 
-    it("403 -> { kind: 'expired' }", async () => {
-      vi.stubGlobal("fetch", vi.fn().mockResolvedValue(emptyResponse(403)));
-      const transport = makeTransport();
-
-      const result = await transport.refreshSession();
-      expect(result.kind).toBe("expired");
-    });
-
-    it("500 -> { kind: 'transient' }", async () => {
+    it("500 -> { kind: 'transient' } (一時障害)", async () => {
       vi.stubGlobal("fetch", vi.fn().mockResolvedValue(emptyResponse(500)));
       const transport = makeTransport();
 
       const result = await transport.refreshSession();
       expect(result.kind).toBe("transient");
     });
-
-    it("network error -> 例外を propagate (controller 側で online retry にハンドリング)", async () => {
-      vi.stubGlobal(
-        "fetch",
-        vi.fn().mockRejectedValue(new TypeError("network")),
-      );
-      const transport = makeTransport();
-
-      await expect(transport.refreshSession()).rejects.toThrow();
-    });
   });
 
   describe("login", () => {
-    it("200 -> session を返し、tokenHolder は更新しない (controller が反映する)", async () => {
+    it("200 + token -> session を返す", async () => {
       vi.stubGlobal(
         "fetch",
         vi.fn().mockResolvedValue(jsonResponse(validSessionBody("login-jwt"))),
       );
-      const tokenHolder = createTokenHolder();
-      const transport = makeTransport({ tokenHolder });
+      const transport = makeTransport();
 
-      const session = await transport.login("u@example.com", "pw");
-
+      const session = await transport.login("user", "pw");
       expect(session.token).toBe("login-jwt");
-      // 重要: transport.login は tokenHolder.setToken を呼ばない (setAccessToken の責務)
-      expect(tokenHolder.getToken()).toBeNull();
     });
 
-    it("401 -> invalidCredentials エラーを throw", async () => {
+    it("401 -> invalidCredentials エラー", async () => {
       vi.stubGlobal("fetch", vi.fn().mockResolvedValue(emptyResponse(401)));
       const transport = makeTransport();
 
-      await expect(transport.login("u", "wrong")).rejects.toThrow(
+      await expect(transport.login("user", "pw")).rejects.toThrow(
         "common:api.invalidCredentials",
       );
     });
 
-    it("500 -> serverError", async () => {
+    it("500 -> serverError エラー", async () => {
       vi.stubGlobal("fetch", vi.fn().mockResolvedValue(emptyResponse(500)));
       const transport = makeTransport();
 
-      await expect(transport.login("u", "p")).rejects.toThrow(
+      await expect(transport.login("user", "pw")).rejects.toThrow(
         "common:api.serverError",
       );
     });
 
-    it("network error -> networkError", async () => {
+    it("network error -> networkError エラー", async () => {
       vi.stubGlobal(
         "fetch",
         vi.fn().mockRejectedValue(new TypeError("network")),
       );
       const transport = makeTransport();
 
-      await expect(transport.login("u", "p")).rejects.toThrow(
+      await expect(transport.login("user", "pw")).rejects.toThrow(
         "common:api.networkError",
       );
-    });
-  });
-
-  describe("logout", () => {
-    it("200 -> { ok: true }", async () => {
-      vi.stubGlobal("fetch", vi.fn().mockResolvedValue(emptyResponse(200)));
-      const transport = makeTransport();
-
-      expect(await transport.logout()).toEqual({ ok: true });
-    });
-
-    it("500 -> { ok: false } (cookie 残存の警告対象)", async () => {
-      vi.stubGlobal("fetch", vi.fn().mockResolvedValue(emptyResponse(500)));
-      const transport = makeTransport();
-
-      expect(await transport.logout()).toEqual({ ok: false });
-    });
-
-    it("network error -> { ok: false }", async () => {
-      vi.stubGlobal(
-        "fetch",
-        vi.fn().mockRejectedValue(new TypeError("network")),
-      );
-      const transport = makeTransport();
-
-      expect(await transport.logout()).toEqual({ ok: false });
-    });
-
-    it("authenticatedFetch 経由で /auth/logout を呼ぶ (Bearer 自動付与 + 401 retry を担う)", async () => {
-      // logout は authenticatedFetch を直接使うので、global fetch ではなく
-      // authenticatedFetch mock 自体に対する呼び出しを検証する
-      const authFetchMock = vi.fn().mockResolvedValue(emptyResponse(200));
-      const transport = makeTransport({ authenticatedFetch: authFetchMock });
-
-      const result = await transport.logout();
-
-      expect(result).toEqual({ ok: true });
-      expect(authFetchMock).toHaveBeenCalledWith(
-        `${apiUrl}/auth/logout`,
-        expect.objectContaining({ method: "POST" }),
-      );
-    });
-
-    it("authenticatedFetch が 401 を返すと { ok: false } (内部で refresh retry も尽きた場合)", async () => {
-      const authFetchMock = vi.fn().mockResolvedValue(emptyResponse(401));
-      const transport = makeTransport({ authenticatedFetch: authFetchMock });
-
-      expect(await transport.logout()).toEqual({ ok: false });
     });
   });
 
@@ -269,7 +153,6 @@ describe("webAuthTransport", () => {
 
       await transport.clearPersistedSession();
 
-      // 副作用がないことを確認 (tokenHolder への影響なし)
       expect(tokenHolder.getToken()).toBe("kept-jwt");
     });
   });
